Created models of USer and Movie all together in the modesl folder
