+++ models/movie.py
@@ -1,0 +1,26 @@
+import uuid
+
+#Relative Import ..2Folders Up, .1Folder up
+#from extensions import db
+
+#absoulut from the base where is the File
+from extensions import db
+class Movie(db.Model):
+    __tablename__ = "movies"
+    id = db.Column(db.String(50), primary_key=True,default=lambda: str(uuid.uuid4()))
+    name = db.Column(db.String(100))
+    poster = db.Column(db.String(255))
+    rating = db.Column(db.Float)
+    summary = db.Column(db.String(500))
+    trailer = db.Column(db.String(255))
+ 
+    # JSON - Keys
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "name": self.name,
+            "poster": self.poster,
+            "rating": self.rating,
+            "summary": self.summary,
+            "trailer": self.trailer,
+        }
